refactor(netx): extract top-level API from netx/httptransport
I put the API there temporarily, waiting for moving the legacy API. Now that
the legacy API has been moved, we can make space for a top-level API.

Part of https://github.com/ooni/probe-engine/issues/746
