+++ cmd/jafar/httpproxy/httpproxy.go
@@ -10,7 +10,7 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/ooni/probe-engine/netx/httptransport"
+	"github.com/ooni/probe-engine/netx"
 )
 
 const product = "jafar/0.1.0"
@@ -27,7 +27,7 @@ type CensoringProxy struct {
 // the Host header of a request. dnsNetwork and dnsAddress are
 // settings to configure the upstream, non censored DNS.
 func NewCensoringProxy(
-	keywords []string, uncensored httptransport.RoundTripper,
+	keywords []string, uncensored netx.HTTPRoundTripper,
 ) *CensoringProxy {
 	return &CensoringProxy{keywords: keywords, transport: uncensored}
 }
