+++ cmd/jafar/resolver/resolver.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 
 	"github.com/miekg/dns"
-	"github.com/ooni/probe-engine/netx/httptransport"
+	"github.com/ooni/probe-engine/netx"
 )
 
 // CensoringResolver is a censoring resolver.
@@ -27,7 +27,7 @@ type CensoringResolver struct {
 // and TLS proxies will pick them up. dnsNetwork and dnsAddress are the
 // settings to configure the upstream, non censored DNS.
 func NewCensoringResolver(
-	blocked, hijacked, ignored []string, uncensored httptransport.Resolver,
+	blocked, hijacked, ignored []string, uncensored netx.Resolver,
 ) *CensoringResolver {
 	return &CensoringResolver{
 		blocked:    blocked,
