+++ cmd/jafar/uncensored/uncensored.go
@@ -10,14 +10,14 @@ import (
 	"github.com/apex/log"
 	"github.com/ooni/probe-engine/experiment/urlgetter"
 	"github.com/ooni/probe-engine/internal/runtimex"
-	"github.com/ooni/probe-engine/netx/httptransport"
+	"github.com/ooni/probe-engine/netx"
 )
 
 // Client is DNS, HTTP, and TCP client.
 type Client struct {
-	dnsClient     *httptransport.DNSClient
-	httpTransport httptransport.RoundTripper
-	dialer        httptransport.Dialer
+	dnsClient     *netx.DNSClient
+	httpTransport netx.HTTPRoundTripper
+	dialer        netx.Dialer
 }
 
 // NewClient creates a new Client.
@@ -33,8 +33,8 @@ func NewClient(resolverURL string) (*Client, error) {
 	}
 	return &Client{
 		dnsClient:     &configuration.DNSClient,
-		httpTransport: httptransport.New(configuration.HTTPConfig),
-		dialer:        httptransport.NewDialer(configuration.HTTPConfig),
+		httpTransport: netx.NewHTTPTransport(configuration.HTTPConfig),
+		dialer:        netx.NewDialer(configuration.HTTPConfig),
 	}, nil
 }
 
@@ -48,39 +48,39 @@ func Must(client *Client, err error) *Client {
 // DefaultClient is the default client for DNS, HTTP, and TCP.
 var DefaultClient = Must(NewClient(""))
 
-var _ httptransport.Resolver = DefaultClient
+var _ netx.Resolver = DefaultClient
 
-// Address implements httptransport.Resolver.Address
+// Address implements netx.Resolver.Address
 func (c *Client) Address() string {
 	return c.dnsClient.Address()
 }
 
-// LookupHost implements httptransport.Resolver.LookupHost
+// LookupHost implements netx.Resolver.LookupHost
 func (c *Client) LookupHost(ctx context.Context, domain string) ([]string, error) {
 	return c.dnsClient.LookupHost(ctx, domain)
 }
 
-// Network implements httptransport.Resolver.Network
+// Network implements netx.Resolver.Network
 func (c *Client) Network() string {
 	return c.dnsClient.Network()
 }
 
-var _ httptransport.Dialer = DefaultClient
+var _ netx.Dialer = DefaultClient
 
-// DialContext implements httptransport.Dialer.DialContext
+// DialContext implements netx.Dialer.DialContext
 func (c *Client) DialContext(ctx context.Context, network, address string) (net.Conn, error) {
 	return c.dialer.DialContext(ctx, network, address)
 }
 
-var _ httptransport.RoundTripper = DefaultClient
+var _ netx.HTTPRoundTripper = DefaultClient
 
-// CloseIdleConnections implement httptransport.RoundTripper.CloseIdleConnections
+// CloseIdleConnections implement netx.HTTPRoundTripper.CloseIdleConnections
 func (c *Client) CloseIdleConnections() {
 	c.dnsClient.CloseIdleConnections()
 	c.httpTransport.CloseIdleConnections()
 }
 
-// RoundTrip implement httptransport.RoundTripper.RoundTrip
+// RoundTrip implement netx.HTTPRoundTripper.RoundTrip
 func (c *Client) RoundTrip(req *http.Request) (*http.Response, error) {
 	return c.httpTransport.RoundTrip(req)
 }
