+++ experiment/hirl/hirl.go
@@ -13,9 +13,9 @@ import (
 
 	"github.com/ooni/probe-engine/internal/randx"
 	"github.com/ooni/probe-engine/model"
+	"github.com/ooni/probe-engine/netx"
 	"github.com/ooni/probe-engine/netx/archival"
 	"github.com/ooni/probe-engine/netx/errorx"
-	"github.com/ooni/probe-engine/netx/httptransport"
 )
 
 const (
@@ -251,7 +251,7 @@ func (meth squidCacheManager) Run(ctx context.Context, config MethodConfig) {
 type RunMethodConfig struct {
 	MethodConfig
 	Name        string
-	NewDialer   func(config httptransport.Config) httptransport.Dialer
+	NewDialer   func(config netx.Config) netx.Dialer
 	RequestLine string
 }
 
@@ -264,9 +264,9 @@ func RunMethod(ctx context.Context, config RunMethodConfig) {
 		config.Out <- result
 	}()
 	if config.NewDialer == nil {
-		config.NewDialer = httptransport.NewDialer
+		config.NewDialer = netx.NewDialer
 	}
-	dialer := config.NewDialer(httptransport.Config{
+	dialer := config.NewDialer(netx.Config{
 		ContextByteCounting: true,
 		Logger:              config.Logger,
 	})
